Refactor all converters tests to call assertProcessedAndWrittenModels util function to do the written, processed and stdout expects.
